Création de la logique pour l ig

+++ app/src/main/java/entity/Task.kt
@@ -1,5 +1,5 @@
 package entity
-data class Task (val id: Long, val description : String, val level : Int, var isDone : Boolean = false)
+data class Task (val id: Int, val description : String, val level : Int, var isDone : Boolean = false)
 {
     fun toggleDone(){
         this.isDone = !this.isDone
